Extract Monitor classes from DDAgentWriter
