+++ dd-trace-ot/src/main/java/datadog/trace/common/writer/Writer.java
@@ -2,6 +2,7 @@ package datadog.trace.common.writer;
 
 import datadog.opentracing.DDSpan;
 import datadog.trace.api.Config;
+import datadog.trace.common.writer.ddagent.Monitor;
 import java.io.Closeable;
 import java.util.List;
 import java.util.Properties;
@@ -65,14 +66,14 @@ public interface Writer extends Closeable {
       return new DDAgentWriter(createApi(config), createMonitor(config));
     }
 
-    private static final DDApi createApi(final Config config) {
+    private static DDApi createApi(final Config config) {
       return new DDApi(
           config.getAgentHost(), config.getAgentPort(), config.getAgentUnixDomainSocket());
     }
 
-    private static final DDAgentWriter.Monitor createMonitor(final Config config) {
+    private static Monitor createMonitor(final Config config) {
       if (!config.isHealthMetricsEnabled()) {
-        return new DDAgentWriter.NoopMonitor();
+        return new Monitor.Noop();
       } else {
         String host = config.getHealthMetricsStatsdHost();
         if (host == null) {
@@ -87,7 +88,7 @@ public interface Writer extends Closeable {
           port = config.getJmxFetchStatsdPort();
         }
 
-        return new DDAgentWriter.StatsDMonitor(host, port);
+        return new Monitor.StatsD(host, port);
       }
     }
